+++ packages/vue/src/components/clipboard/index.ts
@@ -36,6 +36,11 @@ export {
   type ClipboardTriggerProps,
   type ClipboardTriggerBaseProps,
 } from './clipboard-trigger.vue'
+export {
+  default as ClipboardValueText,
+  type ClipboardValueTextProps,
+  type ClipboardValueTextBaseProps,
+} from './clipboard-value-text.vue'
 export { useClipboard, type UseClipboardProps, type UseClipboardReturn } from './use-clipboard'
 export { useClipboardContext, type UseClipboardContext } from './use-clipboard-context'
 export { clipboardAnatomy } from './clipboard.anatomy'